Initial stab at Python3 compatibility with 2to3, then manual changes

+++ others/py3compat.py
@@ -1,7 +1,7 @@
 
 # Workaround for Python3 compat
 try:
-  INTEGER_TYPES = (int, long)
+  INTEGER_TYPES = (int, int)
 except NameError:
   long = int
   INTEGER_TYPES = (int,)
